requirements.txt
feedparser>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
groq>=0.9.0
Pillow>=10.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
cryptography>=41.0.0
lxml>=4.9.0
